Refactor process module testutils
Break apart the process module test utils and some minor renaming.
