+++ src/process/testutil/mod.rs
@@ -1,0 +1,9 @@
+mod assert_process_result;
+mod handle_event_test;
+mod process_module_test;
+
+pub use self::{
+	assert_process_result::_assert_process_result,
+	handle_event_test::handle_event_test,
+	process_module_test::process_module_test,
+};
